+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+## [Unreleased]
+
 ## [4.2.3] - 2017-10-10
 
 * Make mbstring extension optional (though Symfony polyfill).